+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 1.20.0 [Unreleased]
 
+### Added
+- Restore support for the deprecated `prefixPreview` and `suffixPreview` query types.
+
 <!-- TODO: add entries for next release -->
 
 ## 1.19.0